% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_blocks.R
\name{create_blocks}
\alias{create_blocks}
\title{create blocs}
\usage{
create_blocks(genetics_map, delta = 0.001)
}
\arguments{
\item{genetics_map}{genetic map phasee}

\item{delta}{delta}
}
\value{
blocks
}
\description{
create blocs
}
