% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_maps.R
\name{download_rutgers_map}
\alias{download_rutgers_map}
\title{download rutgers maps using the following url : http://compgen.rutgers.edu/downloads/rutgers_map_v3.zip}
\usage{
download_rutgers_map()
}
\value{
None
}
\description{
download rutgers maps using the following url : http://compgen.rutgers.edu/downloads/rutgers_map_v3.zip
}
