% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map_update.R
\name{genetic_map_interp}
\alias{genetic_map_interp}
\title{use and interpolate of all chromosomes using a reference map  object}
\usage{
genetic_map_interp(genetic_map_all, snp_list)
}
\arguments{
\item{genetic_map_all}{reference genetic map of all chromosomes [chr,BP,cM]}

\item{snp_list}{SNP position list to interpolate in all chromosomes [chr,BP]}
}
\value{
gen_map_allchr_interp : genetic map for the SNP list : [chr,BP,cM]
}
\description{
use and interpolate of all chromosomes using a reference map  object
}
