% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map_update.R
\name{genetic_map_interp_chr}
\alias{genetic_map_interp_chr}
\title{use and interpolate genetic map of a chromosome}
\usage{
genetic_map_interp_chr(genetic_map, snp_list)
}
\arguments{
\item{genetic_map}{reference genetic map of a chromosome [BP,cM]}

\item{snp_list}{SNP position list to interpolate [BP]}
}
\value{
gen_map_chr_interp : genetic map for the SNP list : [BP,cM]
}
\description{
use and interpolate genetic map of a chromosome
}
