% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_maps.R
\name{get_rutgers_genMap_SexAverage}
\alias{get_rutgers_genMap_SexAverage}
\title{get rutgers map in genMap format [bp,cM]}
\usage{
get_rutgers_genMap_SexAverage(genetic_map_path, chr)
}
\arguments{
\item{file_path}{the location of the rutgers map}
}
\value{
genetic_map : the genetetic mapin genMap format
}
\description{
get rutgers map in genMap format [bp,cM]
}
