% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics_maps.R
\name{get_rutgers_map}
\alias{get_rutgers_map}
\title{get rutgers map}
\usage{
get_rutgers_map(file_path)
}
\arguments{
\item{file_path}{the location of the rutgers map}
}
\value{
the rutgers map in a data.table structure
}
\description{
get rutgers map
}
