% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{karyotype_plot}
\alias{karyotype_plot}
\title{Genome coverage of haplotype blocs}
\usage{
karyotype_plot(df_blocs, colors = NULL)
}
\arguments{
\item{df_blocs}{data.frame}

\item{colors}{colors (int)}
}
\value{
karyotype plot object
}
\description{
Genome coverage of haplotype blocs
}
