% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map_update.R
\name{read.augmented_genetic_map}
\alias{read.augmented_genetic_map}
\title{update the genetic map}
\usage{
read.augmented_genetic_map(genetic_map_dir, snp_physical_positions)
}
\arguments{
\item{genetic_map_path}{i.e. rutgers map}

\item{to_be_updated_genetic_map_path}{the genetic map that you want to update}

\item{delim}{character used to separate fields within a record. By default tab separated values is used}
}
\value{
a genetic map updated
}
\description{
update the genetic map
}
