% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic_map_update.R
\name{read.genetic_map}
\alias{read.genetic_map}
\title{load the rutgers map to a gen_map dataframe [chr,rsid,BP,cM]}
\usage{
read.genetic_map(genetic_map_dir, chromosomes = 1:23)
}
\arguments{
\item{genetic_map_dir}{path to rutgers map}

\item{chromosome}{the list of chomosomee to load update}

\item{delim}{character used to separate fields within a record. By default tab separated values is used}
}
\value{
a genetic map
}
\description{
load the rutgers map to a gen_map dataframe [chr,rsid,BP,cM]
}
