% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{save_plot}
\alias{save_plot}
\title{Save plot}
\usage{
save_plot(plot, output)
}
\arguments{
\item{plot}{the ggplot object}

\item{output}{the path indicating where to save the plot}
}
\value{
None
}
\description{
Save plot
}
