% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{haplotype_block_distribution}
\alias{haplotype_block_distribution}
\title{haplotype block distribution}
\usage{
haplotype_block_distribution()
}
\value{
ggplot
}
\description{
haplotype block distribution
}
