% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{phenotype_distribution}
\alias{phenotype_distribution}
\title{phenotype distribution}
\usage{
phenotype_distribution(phenotypes_path)
}
\arguments{
\item{phenotypes_path}{list of phenotypes path}
}
\value{
ggplot
}
\description{
phenotype distribution
}
